// RickandMorty-app/src/enums/CharacterEnums.ts
export enum Status {
  Alive = 'Alive',
  Dead = 'Dead',
  Unknown = 'Unknown',
}

export enum Species {
  Human = 'Human',
  Alien = 'Alien',
  Meeseeks = "Meeseeks",
  Cronenberg = "Cronenberg",
  SpermPeople = "Sperm People",
  Gromflomites = "Gromflomites",
  Plumbuses = "Plumbuses",
  Gazorpazorp = "Gazorpazorp",
  Glapflap = "Glapflap",
  KrombopulosMichael = "Krombopulos Michael",
  Birdperson = "Birdperson",
  Shmoglons = "Shmoglons",
  FroopylandCreatures = "Froopyland Creatures",
  ScaryTerry = "Scary Terry",
  Sludgers = "Sludgers",
  Zigerions = "Zigerions",
}

export enum Gender {
  Male = 'Male',
  Female = 'Female',
  Unknown = 'Unknown',
}

export enum Origin {
  EarthDimensionC137 = "Earth (Dimension C-137)",
  CitadelOfRicks = "Citadel of Ricks (Multiverse)",
  GalacticFederation = "Galactic Federation",
  Unknown = "Unknown",
  Gazorpazorpfield = "Gazorpazorpfield (Dimension C-137)",
  Birdworld = "Birdworld (Dimension C-137)",
  Froopyland = "Froopyland (Dimension C-137)",
  NightmareDimension = "Nightmare Dimension",
  ZigerionPlanet = "Zigerion Planet (Dimension C-137)",
}